Refactored some code so that modulesService has less side effects. The new controller still needs to be called from javascript now.
